move methods from S3 to S3Bucket
