+++ Makefile
@@ -10,6 +10,10 @@ test:
 	aws --endpoint-url=${S3_ENDPOINT_URL} s3 rb s3://test || true
 	aws --endpoint-url=${S3_ENDPOINT_URL} s3 mb s3://test
 
+	aws --endpoint-url=${S3_ENDPOINT_URL} s3 rm --recursive s3://versioning-test || true
+	aws --endpoint-url=${S3_ENDPOINT_URL} s3 rb s3://versioning-test || true
+	aws --endpoint-url=${S3_ENDPOINT_URL} s3 mb s3://versioning-test
+
 	aws --endpoint-url=${S3_ENDPOINT_URL} s3 rm --recursive s3://test.bucket || true
 	aws --endpoint-url=${S3_ENDPOINT_URL} s3 rb s3://test.bucket || true
