+++ src/client_test.ts
@@ -1,9 +1,4 @@
-import {
-  assert,
-  assertEquals,
-  assertObjectMatch,
-  assertThrowsAsync,
-} from "../test_deps.ts";
+import { assert, assertEquals, assertThrowsAsync } from "../test_deps.ts";
 import { S3Error } from "./error.ts";
 import { S3Bucket } from "./bucket.ts";
 import { S3 } from "./client.ts";
@@ -76,101 +71,5 @@ Deno.test({
       "creationDate is not of type Date",
     );
     assertEquals(owner.displayName, "minio");
-  },
-});
-
-Deno.test({
-  name: "[client] should put a bucket versioning configuration",
-  async fn() {
-    await s3.putBucketVersioning("test.bucket", { status: "Enabled" });
-
-    let resp = await s3.getBucketVersioning("test.bucket");
-    assertEquals(resp, { status: "Enabled" });
-
-    // teardown
-    await s3.putBucketVersioning("test.bucket", { status: "Suspended" });
-
-    resp = await s3.getBucketVersioning("test.bucket");
-    assertEquals(resp, { status: "Suspended" });
-  },
-});
-
-Deno.test({
-  name: "[client] should list object versions",
-  async fn() {
-    await s3.putBucketVersioning("test.bucket", { status: "Enabled" });
-    const bucket = s3.getBucket("test.bucket");
-
-    await bucket.putObject("test", encoder.encode("test1"));
-    await bucket.deleteObject("test");
-    await bucket.putObject("test", encoder.encode("test1"));
-
-    const resp = await s3.listObjectVersions("test.bucket");
-
-    assertObjectMatch(resp, {
-      delimiter: undefined,
-      encodingType: undefined,
-      isTruncated: false,
-      keyMarker: undefined,
-      maxKeys: 1000,
-      name: "test.bucket",
-      nextKeyMarker: undefined,
-      nextVersionIdMarker: undefined,
-      prefix: undefined,
-      versionIdMarker: undefined,
-    });
-
-    assert(resp.versions);
-    assert(resp.deleteMarkers);
-    assertEquals(resp.versions.length, 2);
-    assertEquals(resp.deleteMarkers.length, 1);
-
-    assertObjectMatch(resp.deleteMarkers[0], {
-      prefix: undefined,
-      isLatest: false,
-      key: "test",
-      owner: {
-        displayName: "minio",
-      },
-    });
-
-    assertObjectMatch(resp.versions[0], {
-      isLatest: true,
-      key: "test",
-      owner: {
-        displayName: "minio",
-      },
-      prefix: undefined,
-      size: 5,
-      storageClass: "STANDARD",
-    });
-
-    assertObjectMatch(resp.versions[1], {
-      isLatest: false,
-      key: "test",
-      owner: {
-        displayName: "minio",
-      },
-      prefix: undefined,
-      size: 5,
-      storageClass: "STANDARD",
-    });
-
-    // teardown
-    await s3.putBucketVersioning("test.bucket", { status: "Suspended" });
-
-    if (resp.versions) {
-      for (const version of resp.versions) {
-        await bucket.deleteObject("test", { versionId: version.versionId });
-      }
-    }
-
-    if (resp.deleteMarkers) {
-      for (const deleteMarker of resp.deleteMarkers) {
-        await bucket.deleteObject("test", {
-          versionId: deleteMarker.versionId,
-        });
-      }
-    }
   },
 });
